Shifting to controller

+++ src/main/java/com/gradproject/yourspace/controller/RequestController.java
@@ -20,21 +20,28 @@ public class RequestController {
 
     @GetMapping("/request/{requestId}")
     public Request getRequest(@PathVariable int requestId) {
+        if (requestService.findById(requestId) == null)
+            throw new RuntimeException("no request found with id " + requestId);
         return requestService.findById(requestId);
     }
 
     @PostMapping("/requests")
     public void saveRequest(@RequestBody Request request) {
+        request.setRequestId(0);
         requestService.saveRequest(request);
     }
 
     @PutMapping("/requests")
     public void updateRequest(@RequestBody Request request) {
+        if (requestService.findById(request.getRequestId()) == null)
+            throw new RuntimeException("no request found with id " + request.getRequestId());
         requestService.updateRequest(request);
     }
 
     @DeleteMapping("/request/{requestId}")
     public void deleteRequest(@PathVariable int requestId) {
+        if (requestService.findById(requestId) == null)
+            throw new RuntimeException("no request found with id " + requestId);
         requestService.deleteRequest(requestId);
     }
 }
